refactor: rename realtime-note-store.service.ts to realtime-note-store.ts
This was committed by mistake, the store should not have the extension `.service`
